Refactored tests to no longer pass true in constructor, the test cases will determine if they want to generate code.  Renamed XJCTest to XJCPrecisionNameTest to better represent what its about.  Added new method to logic to support working with the model before the JCodeModel
