+++ jaxb-ri/xjc/src/test/java/xjcTests/XJCMixedTest.java
@@ -8,7 +8,7 @@ public class XJCMixedTest extends AbstractXJCTest {
 
 	@Test
 	public void simpleMixedTest() throws Throwable {
-		runTest(new Logic(false) {
+		runTest(new Logic() {
 
 			@Override
 			protected File getXsd() {
